+++ src/components/Homepage/HomeCards.tsx
@@ -1,36 +1,20 @@
-import { useGetUser } from "@/services/getUsers"
+import { useGetUser, useGetComments, updateVotes } from "@/services/query"
 import { UserTypes, Comments } from "@/types/types"
 import { ArrowUp, MessageSquare, SendHorizontal, X } from "lucide-react"
 import { useState } from 'react'
-import axios from "axios"
-import { plusVotes, downVotes } from "@/services/resources"
 import {Collapsible, CollapsibleContent, CollapsibleTrigger} from '@/components/ui/collapsible'
 
 export default function HomeCards() {
 
-  const { isLoading, data } = useGetUser();
   const [vote, setVote] = useState(null);
   const [clicked, setClicked] = useState(false);
-  
 
-  const updateVotes = async (user: UserTypes<string, number, boolean>) => {
-    setClicked(!clicked);
-    
-    if (clicked) {
-      user.state = !user.state;
-      return axios.patch(`http://localhost:3000/users/${user.id}/?posts=upvote`, downVotes(user))
-      .then(res => setVote( res.data.posts.downvote))
-      .catch(err => console.log(err));
+  const { isLoading: userLoad, data: userData } = useGetUser();
+  const { isLoading: commentsLoad, data: commentsData, error: commentsError } = useGetComments()
 
-    } else {
-      user.state = !user.state;
-        return axios.patch(`http://localhost:3000/users/${user.id}/?posts=upvote`, plusVotes(user))
-       .then(res => setVote(res.data.posts.upvote))
-       .catch(err => console.log(err));
-    }
-  }
+  console.log(commentsData?.data)
 
-  if (isLoading) {
+  if (userLoad) {
     return <h1>Loading ... </h1>
   }
 
@@ -38,7 +22,7 @@ export default function HomeCards() {
   return (
     
     <div className='flex flex-col space-y-10 mt-10'>      
-        {data?.data.map((user: UserTypes<string, number, boolean>) => {
+        {userData?.data.map((user: UserTypes<string, number, boolean>) => {
 
             return (
               <Collapsible key={user.id} className={`border rounded-md border-cyan-300 max-w-xl mx-auto p-5 shadow-md`}>
@@ -61,7 +45,10 @@ export default function HomeCards() {
                 <p className='mt-5'>{user.posts.summary}</p>
 
                 <div className='flex gap-5 mt-2'>
-                  <section className='flex gap-1 hover: cursor-pointer' onClick={() => updateVotes(user)}>        
+                  <section className='flex gap-1 hover: cursor-pointer' onClick={() => {
+                    setClicked(!clicked)
+                    updateVotes(user, clicked, setClicked)
+                  }}>        
 
                     <ArrowUp size={18} className="mt-1" color={user.state ? user.color : user.unclicked}/>
